+++ lib/index.ts
@@ -63,6 +63,7 @@ export * from './npc-generation/socialClass'
 export * from './npc-generation/createName'
 export * from './npc-generation/createPersonality'
 export * from './npc-generation/familyRaces'
+export * from './npc-generation/familyRelationships'
 export * from './npc-generation/fetchProfessionChance'
 export * from './npc-generation/fetchRace'
 export * from './npc-generation/filterNpcByProfession'
